Issue 10 - Partial fix for hiding More Results when there are no more hits
This makes sure that the button is not displayed if the number of hits is
lower than the number that is to be displayed on the first page of hits.
(3 of 3 hits)

TODO: Fix it when we have more than one page of hits, but have displayed
all of them.